Revert "ES-6566: Move the calculation of data tier usage stats to individual nodes (#100230) (#101599)" (#102042)
Reverting because the new action is not properly handled in a mixed
cluster.
