--- docs/changelog/101599.yaml
@@ -1,6 +1,0 @@
-pr: 101599
-summary: Move the calculation of data tier usage stats to individual nodes
-area: ILM+SLM
-type: bug
-issues:
-  - 100230
